requirements.txt
flask
requests
beautifulsoup4
feedparser
python-dotenv
gunicorn